Move utils back to /api
